01-tutorials/workshop/session1/pyproject.toml
[project]
name = "session1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.3",
    "langchain-openai>=1.1.3",
    "python-dotenv>=1.2.1",
    "agentkit-sdk-python>=0.2.0"
]
